requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
boto3==1.42.19
botocore==1.42.19
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
dill==0.4.0
dnspython==2.8.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
fonttools==4.61.1
from-root==1.3.0
h11==0.16.0
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mypy-boto3-s3==1.42.16
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==2.2.6
-e git+https://github.com/rahultembhurney/one_more_project.git@59e0bc183fa06d3401495f6b2e68f65a608415a2#egg=one_more_project
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.5
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
stack-data==0.6.3
starlette==0.50.0
threadpoolctl==3.6.0
tomli==2.3.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
